Add files via upload
This code only connects the source point to the closest point in Manhattan distance without ties. Reference means point the line will start from. Potential means the point the line will end at. Need to dynamically allocate the arrays in lines 77, 78, 82, and 83 to take in a variable number of points, handle ties in Manhattan distance, and take existing for loop to create more than 1 line.